Renamed domain from chat app to chatterix
