+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.Chatterix"
         tools:targetApi="34">
         <activity
-            android:name="com.huzil.chatapp.presentation.MainActivity"
+            android:name="com.huzil.chatterix.presentation.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Chatterix">
             <intent-filter>
