--- app/src/main/java/com/huzil/chatapp/di/FirebaseModule.kt
@@ -1,17 +1,0 @@
-package com.huzil.chatapp.di
-
-import com.google.firebase.auth.FirebaseAuth
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object FirebaseModule {
-
-    @Singleton
-    @Provides
-    fun provideAuthentication() = FirebaseAuth.getInstance()
-}
