--- app/src/main/java/com/huzil/chatapp/domain/repository/AuthenticationRepository.kt
@@ -1,5 +1,0 @@
-package com.huzil.chatapp.domain.repository
-
-interface AuthenticationRepository {
-    suspend fun register(email: String, password: String)
-}
