--- app/src/main/java/com/huzil/chatapp/presentation/MainActivity.kt
@@ -1,19 +1,0 @@
-package com.huzil.chatapp.presentation
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import com.huzil.chatapp.presentation.ui.theme.ChatterixTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            ChatterixTheme {
-
-            }
-        }
-    }
-}
