--- app/src/main/java/com/huzil/chatapp/presentation/authentication/login/LoginViewModel.kt
@@ -1,21 +1,0 @@
-package com.huzil.chatapp.presentation.authentication.login
-
-import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-
-@HiltViewModel
-class LoginViewModel @Inject constructor(
-    savedStateHandle: SavedStateHandle
-) : ViewModel() {
-
-    private val _stateFlow: MutableStateFlow<LoginState> = MutableStateFlow(LoginState())
-
-    val stateFlow: StateFlow<LoginState> = _stateFlow.asStateFlow()
-
-
-}
