--- app/src/main/java/com/huzil/chatapp/presentation/authentication/register/RegisterContract.kt
@@ -1,31 +1,0 @@
-package com.huzil.chatapp.presentation.authentication.register
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-
-/**
- * UI State that represents RegisterScreen
- **/
-data class RegisterState(
-    val text: String = ""
-)
-
-/**
- * Register Actions emitted from the UI Layer
- * passed to the coordinator to handle
- **/
-data class RegisterActions(
-    val onClick: () -> Unit = {}
-)
-
-
-/**
- * PreviewParameter Provider for RegisterScreen Preview
- * Add values to the sequence to see the preview in different states
- **/
-class RegisterStatePreviewParameterProvider : PreviewParameterProvider<RegisterState> {
-    override val values: Sequence<RegisterState>
-        get() = sequenceOf(
-
-        )
-}
