--- app/src/main/java/com/huzil/chatapp/presentation/authentication/register/RegisterViewModel.kt
@@ -1,21 +1,0 @@
-package com.huzil.chatapp.presentation.authentication.register
-
-import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-
-@HiltViewModel
-class RegisterViewModel @Inject constructor(
-    savedStateHandle: SavedStateHandle
-) : ViewModel() {
-
-    private val _stateFlow: MutableStateFlow<RegisterState> = MutableStateFlow(RegisterState())
-
-    val stateFlow: StateFlow<RegisterState> = _stateFlow.asStateFlow()
-
-
-}
