+++ app/src/main/java/com/huzil/chatterix/domain/repository/AuthenticationRepository.kt
@@ -1,0 +1,5 @@
+package com.huzil.chatterix.domain.repository
+
+interface AuthenticationRepository {
+    suspend fun register(email: String, password: String)
+}
