+++ app/src/main/java/com/huzil/chatterix/presentation/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.huzil.chatterix.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.huzil.chatterix.presentation.ui.theme.ChatterixTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            ChatterixTheme {
+
+            }
+        }
+    }
+}
