+++ app/src/main/java/com/huzil/chatterix/presentation/authentication/login/LoginCoordinator.kt
@@ -1,0 +1,30 @@
+package com.huzil.chatterix.presentation.authentication.login
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.hilt.navigation.compose.hiltViewModel
+
+/**
+ * Screen's coordinator which is responsible for handling actions from the UI layer
+ * and one-shot actions based on the new UI state
+ */
+class LoginCoordinator(
+    val viewModel: LoginViewModel
+) {
+    val screenStateFlow = viewModel.stateFlow
+
+    fun doStuff() {
+        // TODO Handle UI Action
+    }
+}
+
+@Composable
+fun rememberLoginCoordinator(
+    viewModel: LoginViewModel = hiltViewModel()
+): LoginCoordinator {
+    return remember(viewModel) {
+        LoginCoordinator(
+            viewModel = viewModel
+        )
+    }
+}
