+++ app/src/main/java/com/huzil/chatterix/presentation/ui/theme/Color.kt
@@ -1,0 +1,225 @@
+package com.huzil.chatterix.presentation.ui.theme
+import androidx.compose.ui.graphics.Color
+
+val primaryLight = Color(0xFF4B6268)
+val onPrimaryLight = Color(0xFFFFFFFF)
+val primaryContainerLight = Color(0xFFD7F0F7)
+val onPrimaryContainerLight = Color(0xFF3A5157)
+val secondaryLight = Color(0xFF5D5F5F)
+val onSecondaryLight = Color(0xFFFFFFFF)
+val secondaryContainerLight = Color(0xFFFFFFFF)
+val onSecondaryContainerLight = Color(0xFF575859)
+val tertiaryLight = Color(0xFF4B6268)
+val onTertiaryLight = Color(0xFFFFFFFF)
+val tertiaryContainerLight = Color(0xFFD7F0F7)
+val onTertiaryContainerLight = Color(0xFF3A5157)
+val errorLight = Color(0xFFBA1A1A)
+val onErrorLight = Color(0xFFFFFFFF)
+val errorContainerLight = Color(0xFFFFDAD6)
+val onErrorContainerLight = Color(0xFF410002)
+val backgroundLight = Color(0xFFFAF9F9)
+val onBackgroundLight = Color(0xFF1B1C1C)
+val surfaceLight = Color(0xFFFCF8F8)
+val onSurfaceLight = Color(0xFF1C1B1C)
+val surfaceVariantLight = Color(0xFFDFE3E4)
+val onSurfaceVariantLight = Color(0xFF434748)
+val outlineLight = Color(0xFF737878)
+val outlineVariantLight = Color(0xFFC3C7C8)
+val scrimLight = Color(0xFF000000)
+val inverseSurfaceLight = Color(0xFF313030)
+val inverseOnSurfaceLight = Color(0xFFF4F0F0)
+val inversePrimaryLight = Color(0xFFB2CBD2)
+val surfaceDimLight = Color(0xFFDDD9D9)
+val surfaceBrightLight = Color(0xFFFCF8F8)
+val surfaceContainerLowestLight = Color(0xFFFFFFFF)
+val surfaceContainerLowLight = Color(0xFFF7F3F2)
+val surfaceContainerLight = Color(0xFFF1EDED)
+val surfaceContainerHighLight = Color(0xFFEBE7E7)
+val surfaceContainerHighestLight = Color(0xFFE5E2E1)
+
+val primaryLightMediumContrast = Color(0xFF30464C)
+val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
+val primaryContainerLightMediumContrast = Color(0xFF61787F)
+val onPrimaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val secondaryLightMediumContrast = Color(0xFF414343)
+val onSecondaryLightMediumContrast = Color(0xFFFFFFFF)
+val secondaryContainerLightMediumContrast = Color(0xFF737575)
+val onSecondaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val tertiaryLightMediumContrast = Color(0xFF30464C)
+val onTertiaryLightMediumContrast = Color(0xFFFFFFFF)
+val tertiaryContainerLightMediumContrast = Color(0xFF61787F)
+val onTertiaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val errorLightMediumContrast = Color(0xFF8C0009)
+val onErrorLightMediumContrast = Color(0xFFFFFFFF)
+val errorContainerLightMediumContrast = Color(0xFFDA342E)
+val onErrorContainerLightMediumContrast = Color(0xFFFFFFFF)
+val backgroundLightMediumContrast = Color(0xFFFAF9F9)
+val onBackgroundLightMediumContrast = Color(0xFF1B1C1C)
+val surfaceLightMediumContrast = Color(0xFFFCF8F8)
+val onSurfaceLightMediumContrast = Color(0xFF1C1B1C)
+val surfaceVariantLightMediumContrast = Color(0xFFDFE3E4)
+val onSurfaceVariantLightMediumContrast = Color(0xFF3F4444)
+val outlineLightMediumContrast = Color(0xFF5B6060)
+val outlineVariantLightMediumContrast = Color(0xFF777B7C)
+val scrimLightMediumContrast = Color(0xFF000000)
+val inverseSurfaceLightMediumContrast = Color(0xFF313030)
+val inverseOnSurfaceLightMediumContrast = Color(0xFFF4F0F0)
+val inversePrimaryLightMediumContrast = Color(0xFFB2CBD2)
+val surfaceDimLightMediumContrast = Color(0xFFDDD9D9)
+val surfaceBrightLightMediumContrast = Color(0xFFFCF8F8)
+val surfaceContainerLowestLightMediumContrast = Color(0xFFFFFFFF)
+val surfaceContainerLowLightMediumContrast = Color(0xFFF7F3F2)
+val surfaceContainerLightMediumContrast = Color(0xFFF1EDED)
+val surfaceContainerHighLightMediumContrast = Color(0xFFEBE7E7)
+val surfaceContainerHighestLightMediumContrast = Color(0xFFE5E2E1)
+
+val primaryLightHighContrast = Color(0xFF0E252B)
+val onPrimaryLightHighContrast = Color(0xFFFFFFFF)
+val primaryContainerLightHighContrast = Color(0xFF30464C)
+val onPrimaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val secondaryLightHighContrast = Color(0xFF202323)
+val onSecondaryLightHighContrast = Color(0xFFFFFFFF)
+val secondaryContainerLightHighContrast = Color(0xFF414343)
+val onSecondaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val tertiaryLightHighContrast = Color(0xFF0E252B)
+val onTertiaryLightHighContrast = Color(0xFFFFFFFF)
+val tertiaryContainerLightHighContrast = Color(0xFF30464C)
+val onTertiaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val errorLightHighContrast = Color(0xFF4E0002)
+val onErrorLightHighContrast = Color(0xFFFFFFFF)
+val errorContainerLightHighContrast = Color(0xFF8C0009)
+val onErrorContainerLightHighContrast = Color(0xFFFFFFFF)
+val backgroundLightHighContrast = Color(0xFFFAF9F9)
+val onBackgroundLightHighContrast = Color(0xFF1B1C1C)
+val surfaceLightHighContrast = Color(0xFFFCF8F8)
+val onSurfaceLightHighContrast = Color(0xFF000000)
+val surfaceVariantLightHighContrast = Color(0xFFDFE3E4)
+val onSurfaceVariantLightHighContrast = Color(0xFF202525)
+val outlineLightHighContrast = Color(0xFF3F4444)
+val outlineVariantLightHighContrast = Color(0xFF3F4444)
+val scrimLightHighContrast = Color(0xFF000000)
+val inverseSurfaceLightHighContrast = Color(0xFF313030)
+val inverseOnSurfaceLightHighContrast = Color(0xFFFFFFFF)
+val inversePrimaryLightHighContrast = Color(0xFFD7F0F7)
+val surfaceDimLightHighContrast = Color(0xFFDDD9D9)
+val surfaceBrightLightHighContrast = Color(0xFFFCF8F8)
+val surfaceContainerLowestLightHighContrast = Color(0xFFFFFFFF)
+val surfaceContainerLowLightHighContrast = Color(0xFFF7F3F2)
+val surfaceContainerLightHighContrast = Color(0xFFF1EDED)
+val surfaceContainerHighLightHighContrast = Color(0xFFEBE7E7)
+val surfaceContainerHighestLightHighContrast = Color(0xFFE5E2E1)
+
+val primaryDark = Color(0xFFFFFFFF)
+val onPrimaryDark = Color(0xFF1D3439)
+val primaryContainerDark = Color(0xFFC0D9E0)
+val onPrimaryContainerDark = Color(0xFF2C4348)
+val secondaryDark = Color(0xFFFFFFFF)
+val onSecondaryDark = Color(0xFF2F3131)
+val secondaryContainerDark = Color(0xFFD4D4D4)
+val onSecondaryContainerDark = Color(0xFF3E4040)
+val tertiaryDark = Color(0xFFFFFFFF)
+val onTertiaryDark = Color(0xFF1D3439)
+val tertiaryContainerDark = Color(0xFFC0D9E0)
+val onTertiaryContainerDark = Color(0xFF2C4348)
+val errorDark = Color(0xFFFFB4AB)
+val onErrorDark = Color(0xFF690005)
+val errorContainerDark = Color(0xFF93000A)
+val onErrorContainerDark = Color(0xFFFFDAD6)
+val backgroundDark = Color(0xFF121414)
+val onBackgroundDark = Color(0xFFE3E2E2)
+val surfaceDark = Color(0xFF141313)
+val onSurfaceDark = Color(0xFFE5E2E1)
+val surfaceVariantDark = Color(0xFF434748)
+val onSurfaceVariantDark = Color(0xFFC3C7C8)
+val outlineDark = Color(0xFF8D9192)
+val outlineVariantDark = Color(0xFF434748)
+val scrimDark = Color(0xFF000000)
+val inverseSurfaceDark = Color(0xFFE5E2E1)
+val inverseOnSurfaceDark = Color(0xFF313030)
+val inversePrimaryDark = Color(0xFF4B6268)
+val surfaceDimDark = Color(0xFF141313)
+val surfaceBrightDark = Color(0xFF3A3939)
+val surfaceContainerLowestDark = Color(0xFF0E0E0E)
+val surfaceContainerLowDark = Color(0xFF1C1B1C)
+val surfaceContainerDark = Color(0xFF201F20)
+val surfaceContainerHighDark = Color(0xFF2A2A2A)
+val surfaceContainerHighestDark = Color(0xFF353435)
+
+val primaryDarkMediumContrast = Color(0xFFFFFFFF)
+val onPrimaryDarkMediumContrast = Color(0xFF1D3439)
+val primaryContainerDarkMediumContrast = Color(0xFFC0D9E0)
+val onPrimaryContainerDarkMediumContrast = Color(0xFF0A2328)
+val secondaryDarkMediumContrast = Color(0xFFFFFFFF)
+val onSecondaryDarkMediumContrast = Color(0xFF2F3131)
+val secondaryContainerDarkMediumContrast = Color(0xFFD4D4D4)
+val onSecondaryContainerDarkMediumContrast = Color(0xFF1E2020)
+val tertiaryDarkMediumContrast = Color(0xFFFFFFFF)
+val onTertiaryDarkMediumContrast = Color(0xFF1D3439)
+val tertiaryContainerDarkMediumContrast = Color(0xFFC0D9E0)
+val onTertiaryContainerDarkMediumContrast = Color(0xFF0A2328)
+val errorDarkMediumContrast = Color(0xFFFFBAB1)
+val onErrorDarkMediumContrast = Color(0xFF370001)
+val errorContainerDarkMediumContrast = Color(0xFFFF5449)
+val onErrorContainerDarkMediumContrast = Color(0xFF000000)
+val backgroundDarkMediumContrast = Color(0xFF121414)
+val onBackgroundDarkMediumContrast = Color(0xFFE3E2E2)
+val surfaceDarkMediumContrast = Color(0xFF141313)
+val onSurfaceDarkMediumContrast = Color(0xFFFEFAF9)
+val surfaceVariantDarkMediumContrast = Color(0xFF434748)
+val onSurfaceVariantDarkMediumContrast = Color(0xFFC8CBCC)
+val outlineDarkMediumContrast = Color(0xFF9FA3A4)
+val outlineVariantDarkMediumContrast = Color(0xFF808484)
+val scrimDarkMediumContrast = Color(0xFF000000)
+val inverseSurfaceDarkMediumContrast = Color(0xFFE5E2E1)
+val inverseOnSurfaceDarkMediumContrast = Color(0xFF2A2A2A)
+val inversePrimaryDarkMediumContrast = Color(0xFF354B51)
+val surfaceDimDarkMediumContrast = Color(0xFF141313)
+val surfaceBrightDarkMediumContrast = Color(0xFF3A3939)
+val surfaceContainerLowestDarkMediumContrast = Color(0xFF0E0E0E)
+val surfaceContainerLowDarkMediumContrast = Color(0xFF1C1B1C)
+val surfaceContainerDarkMediumContrast = Color(0xFF201F20)
+val surfaceContainerHighDarkMediumContrast = Color(0xFF2A2A2A)
+val surfaceContainerHighestDarkMediumContrast = Color(0xFF353435)
+
+val primaryDarkHighContrast = Color(0xFFFFFFFF)
+val onPrimaryDarkHighContrast = Color(0xFF000000)
+val primaryContainerDarkHighContrast = Color(0xFFC0D9E0)
+val onPrimaryContainerDarkHighContrast = Color(0xFF000000)
+val secondaryDarkHighContrast = Color(0xFFFFFFFF)
+val onSecondaryDarkHighContrast = Color(0xFF000000)
+val secondaryContainerDarkHighContrast = Color(0xFFD4D4D4)
+val onSecondaryContainerDarkHighContrast = Color(0xFF000000)
+val tertiaryDarkHighContrast = Color(0xFFFFFFFF)
+val onTertiaryDarkHighContrast = Color(0xFF000000)
+val tertiaryContainerDarkHighContrast = Color(0xFFC0D9E0)
+val onTertiaryContainerDarkHighContrast = Color(0xFF000000)
+val errorDarkHighContrast = Color(0xFFFFF9F9)
+val onErrorDarkHighContrast = Color(0xFF000000)
+val errorContainerDarkHighContrast = Color(0xFFFFBAB1)
+val onErrorContainerDarkHighContrast = Color(0xFF000000)
+val backgroundDarkHighContrast = Color(0xFF121414)
+val onBackgroundDarkHighContrast = Color(0xFFE3E2E2)
+val surfaceDarkHighContrast = Color(0xFF141313)
+val onSurfaceDarkHighContrast = Color(0xFFFFFFFF)
+val surfaceVariantDarkHighContrast = Color(0xFF434748)
+val onSurfaceVariantDarkHighContrast = Color(0xFFF8FBFC)
+val outlineDarkHighContrast = Color(0xFFC8CBCC)
+val outlineVariantDarkHighContrast = Color(0xFFC8CBCC)
+val scrimDarkHighContrast = Color(0xFF000000)
+val inverseSurfaceDarkHighContrast = Color(0xFFE5E2E1)
+val inverseOnSurfaceDarkHighContrast = Color(0xFF000000)
+val inversePrimaryDarkHighContrast = Color(0xFF162D33)
+val surfaceDimDarkHighContrast = Color(0xFF141313)
+val surfaceBrightDarkHighContrast = Color(0xFF3A3939)
+val surfaceContainerLowestDarkHighContrast = Color(0xFF0E0E0E)
+val surfaceContainerLowDarkHighContrast = Color(0xFF1C1B1C)
+val surfaceContainerDarkHighContrast = Color(0xFF201F20)
+val surfaceContainerHighDarkHighContrast = Color(0xFF2A2A2A)
+val surfaceContainerHighestDarkHighContrast = Color(0xFF353435)
+
+
+
+
+
+
+
